+++ client/src/hooks/Input/index.ts
@@ -2,3 +2,4 @@ export { default as useUserKey } from './useUserKey';
 export { default as useDebounce } from './useDebounce';
 export { default as useTextarea } from './useTextarea';
 export { default as useRequiresKey } from './useRequiresKey';
+export { default as useMultipleKeys } from './useMultipleKeys';
